Bug 1942839: Fix async return
