+++ frontend/packages/kubevirt-plugin/src/components/vm-disks/disk-summary.tsx
@@ -1,12 +1,13 @@
-import * as React from 'react';
 import * as _ from 'lodash';
+import * as React from 'react';
 import { useTranslation } from 'react-i18next';
+
+import { winToolsContainerNames } from '../../constants/vm/wintools';
 import {
   getContainerImageByDisk,
-  getURLSourceByDisk,
   getPVCSourceByDisk,
+  getURLSourceByDisk,
 } from '../../selectors/vm/selectors';
-import { WINTOOLS_CONTAINER_NAMES } from '../../constants';
 import { VMKind } from '../../types';
 import { V1Disk } from '../../types/vm/disk/V1Disk';
 
@@ -23,8 +24,8 @@ export const DiskSummary: React.FC<DiskSummaryProps> = ({ disks, vm }) => {
         const url = getURLSourceByDisk(vm, name);
         const nameKey = `kubevirt-disk-summary-disk-title-${name}`;
         let value = '';
-
-        if (_.includes(WINTOOLS_CONTAINER_NAMES, container)) {
+        const containerNames = winToolsContainerNames();
+        if (_.includes(containerNames, container)) {
           value = t('kubevirt-plugin~Windows tools: {{container}}', { container });
         } else if (container) {
           value = t('kubevirt-plugin~Container: {{container}}', { container });
